requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-restful==0.5.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
mypy-extensions==1.0.0
psutil==5.9.5
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.23.2
